+++ packages/cohook-preact/CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.2.0 (2021-10-24)
+
+
+### ✅ Tests | 测试
+
+* add test case ([e6f485e](https://github.com/Keylenn/cohookjs/commit/e6f485e7c89e7dac2958012f24c0c8562598ba5d))
+
+
+### 🚀 Chore | 构建/工程依赖/工具
+
+* add .gitignore ([5ef92ca](https://github.com/Keylenn/cohookjs/commit/5ef92ca43327bf7bf6d43e86758d88a3e95d2d2a))
+* eslint ([4d56e5f](https://github.com/Keylenn/cohookjs/commit/4d56e5f0a0b1a99bb787e9152c9dae06169682fe))
+* fix husky-v7 config ([b1630d6](https://github.com/Keylenn/cohookjs/commit/b1630d6b1887d3ca29579e54319cd2818d39ba1e))
+* **release:** 1.1.0-beta.0 ([aaf00e5](https://github.com/Keylenn/cohookjs/commit/aaf00e5b6c3be8d2dffd4cc3b4c1cf267900e665))
+* **release:** 1.1.0-beta.1 ([df37baf](https://github.com/Keylenn/cohookjs/commit/df37baf8155b5bf67272724a00c8f9d255004700))
+
+
+### 🐛 Bug Fixes | Bug 修复
+
+* add index ([8313dd3](https://github.com/Keylenn/cohookjs/commit/8313dd341fc51843b60bcfd7e11d99fe0155b86f))
+* build ([316198e](https://github.com/Keylenn/cohookjs/commit/316198e33e39ec0fc31c8292787865d842801c08))
+* change immer export way ([faa3f5f](https://github.com/Keylenn/cohookjs/commit/faa3f5fb8e3c1bcf033f26b941d602d3604bc73a))
+* init ([9b46338](https://github.com/Keylenn/cohookjs/commit/9b4633825a8b18e066ab1ba5f09ef85ec4e0ab60))
+* reset packages ([ab75158](https://github.com/Keylenn/cohookjs/commit/ab75158999acbca6c582a85f3abc23f7bc39e546))
+
+
+### ✨ Features | 新功能
+
+* add share option ([2560543](https://github.com/Keylenn/cohookjs/commit/2560543fa842d5d74650e7bc560b25f15ef6adcb))
+* init ([f02fb79](https://github.com/Keylenn/cohookjs/commit/f02fb7997d39265c7b5bc03cdbddab3517cfa257))
+* init ([16ba0f7](https://github.com/Keylenn/cohookjs/commit/16ba0f760a09ffd178b49b6fe4187484590dc98c))
+* init ([5da9f01](https://github.com/Keylenn/cohookjs/commit/5da9f0120f88c62600c66029941a80cb4122d20c))
+* init ([3121946](https://github.com/Keylenn/cohookjs/commit/312194675379894f80917d4536cef264d7b9e8a4))
+
 ## [1.1.0-beta.1](https://github.com/Keylenn/cohookjs/compare/v1.1.0-beta.0...v1.1.0-beta.1) (2021-09-26)